Try adding PSR-4 style autoloading
For #3
